requirements.txt
azure-cosmos
azure-storage-blob
azure-functions
